@@ -1,17 +1,15 @@
 
 plugins {
-    id("org.springframework.boot") version "2.7.10"
-    id("io.spring.dependency-management") version "1.0.15.RELEASE"
+    id("org.springframework.boot") version "3.1.0"
+    id("io.spring.dependency-management") version "1.1.0"
     id("java")
 }
 
 group = "com.vmware.tap.accelerators"
 version = "0.0.1-SNAPSHOT"
-java.sourceCompatibility = JavaVersion.VERSION_11
+java.sourceCompatibility = JavaVersion.VERSION_17
 
-val springdocVersion = "1.6.14"
-
-extra["snakeyaml.version"] = "1.33"
+val springdocVersion = "2.1.0"
 
 repositories {
     mavenCentral()
@@ -23,8 +21,9 @@ dependencies {
     implementation("org.springframework.boot:spring-boot-starter-data-jpa")
     implementation("org.springframework.boot:spring-boot-starter-web")
     implementation("org.springframework.boot:spring-boot-starter-validation")
+    developmentOnly("org.springframework.boot:spring-boot-docker-compose")
 
-    implementation("org.springdoc:springdoc-openapi-webmvc-core:${springdocVersion}")
+    implementation("org.springdoc:springdoc-openapi-starter-webmvc-api:${springdocVersion}")
 
     // Infrastructure
     implementation("org.flywaydb:flyway-core")
